+++ demo/pages/use-with-react.vue
@@ -77,7 +77,7 @@ watch(locale, value => {
           <pre>
 &lt;add-to-calendar-button
   name="{{ $t('demo_data.name_dummy') }}"
-  options="'Apple','Google'"
+  options="'apple','google'"
   location="{{ $t('demo_data.location') }}"
   startDate="{{ defaultDate }}"
   endDate="{{ defaultDate }}"
@@ -93,7 +93,7 @@ watch(locale, value => {
           <pre>
 &lt;AddToCalendarButton
   name="{{ $t('demo_data.name_dummy') }}"
-  options={['Apple','Google']}
+  options={['apple','google']}
   location="{{ $t('demo_data.location') }}"
   startDate="{{ defaultDate }}"
   endDate="{{ defaultDate }}"