+++ demo/pages/use-with-svelte.vue
@@ -50,7 +50,7 @@ watch(locale, value => {
           <pre>
 &lt;add-to-calendar-button
   name="{{ $t('demo_data.name_dummy') }}"
-  options="'Apple','Google'"
+  options="'apple','google'"
   location="{{ $t('demo_data.location') }}"
   startDate="{{ defaultDate }}"
   endDate="{{ defaultDate }}"