+++ demo/pages/use-with-vue.vue
@@ -81,7 +81,7 @@ compilerOptions: {
           <pre>
 &lt;add-to-calendar-button
   name="{{ $t('demo_data.name_dummy') }}"
-  options="'Apple','Google'"
+  options="'apple','google'"
   location="{{ $t('demo_data.location') }}"
   startDate="{{ defaultDate }}"
   endDate="{{ defaultDate }}"