+++ src/generators/rich-data.ts
@@ -34,7 +34,7 @@ function atcb_generate_rich_data(data: ATCBConfig, parent: Element): void {
     if (data.dates!.length > 1) {
       schemaContent.push('"@id":' + JSON.stringify(id + '-' + (i + 1)));
     }
-    if (data.dates![`${i}`]!.status!.toLowerCase() === 'cancelled') {
+    if (data.dates![`${i}`]!.status === 'cancelled') {
       schemaContent.push('"eventStatus":"https://schema.org/EventCancelled"');
     } else {
       schemaContent.push('"eventStatus":"https://schema.org/EventScheduled"');